Refactor code to align with new SDK recommendations
* Split Activities and Workflows into separate files
* Remove test fixtures to reduce complexity for intro tutorial.
