+++ .gitignore
@@ -1,4 +1,5 @@
 .pytest_cache
 __pycache__
 venv
+env
