+++ activities.py
@@ -1,0 +1,8 @@
+# @@@SNIPSTART python-project-template-activities
+from temporalio import activity
+
+@activity.defn
+async def say_hello(name: str) -> str:
+    return f"Hello, {name}!"
+
+# @@@SNIPEND
