+++ tests/test_run_worker.py
@@ -1,30 +1,32 @@
 # @@@SNIPSTART hello-world-project-template-python-tests
-import asyncio
 import uuid
 
 import pytest
-from run_worker import SayHello, say_hello
+
+from activities import say_hello
+from run_worker import SayHello
 from temporalio import activity
-from temporalio.client import Client
 from temporalio.worker import Worker
+from temporalio.testing import WorkflowEnvironment
 
 
 @pytest.mark.asyncio
-async def test_execute_workflow(client: Client):
+async def test_execute_workflow():
     task_queue_name = str(uuid.uuid4())
+    async with await WorkflowEnvironment.start_local() as env:
 
-    async with Worker(
-        client,
-        task_queue=task_queue_name,
-        workflows=[SayHello],
-        activities=[say_hello],
-    ):
-        assert "Hello, World!" == await client.execute_workflow(
-            SayHello.run,
-            "World",
-            id=str(uuid.uuid4()),
+        async with Worker(
+            env.client,
             task_queue=task_queue_name,
-        )
+            workflows=[SayHello],
+            activities=[say_hello],
+        ):
+            assert "Hello, World!" == await env.client.execute_workflow(
+                SayHello.run,
+                "World",
+                id=str(uuid.uuid4()),
+                task_queue=task_queue_name,
+            )
 
 
 @activity.defn(name="say_hello")
@@ -33,20 +35,20 @@ async def say_hello_mocked(name: str) -> str:
 
 
 @pytest.mark.asyncio
-async def test_mock_activity(client: Client):
+async def test_mock_activity():
     task_queue_name = str(uuid.uuid4())
-    async with Worker(
-        client,
-        task_queue=task_queue_name,
-        workflows=[SayHello],
-        activities=[say_hello_mocked],
-    ):
-        assert "Hello, World from mocked activity!" == await client.execute_workflow(
-            SayHello.run,
-            "World",
-            id=str(uuid.uuid4()),
+    async with await WorkflowEnvironment.start_local() as env:
+        async with Worker(
+            env.client,
             task_queue=task_queue_name,
-        )
-
+            workflows=[SayHello],
+            activities=[say_hello_mocked],
+        ):
+            assert "Hello, World from mocked activity!" == await env.client.execute_workflow(
+                SayHello.run,
+                "World",
+                id=str(uuid.uuid4()),
+                task_queue=task_queue_name,
+            )
 
 # @@@SNIPEND
